# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "船舶能效分析平台 API",
    "description": "整合 Meta / Identity / Vessel / Data 四个微服务的 OpenAPI 文档。\n\n每条路径通过 path-level `servers` 字段标明所属服务：\n\n| 服务 | 地址 |\n|------|------|\n| Meta 元数据 | http://localhost:8000 |\n| Identity 身份认证 | http://localhost:8001 |\n| Vessel 船舶管理 | http://localhost:8002 |\n| Data 遥测数据 | http://localhost:8003 |",
    "version": "0.1.0"
  },
  "paths": {
...
    "/upload/vessel/{vessel_id}/standard": {
      "post": {
        "summary": "上传标准 CSV 数据",
        "description": "接收船舶 CSV 文件，立即返回 202，后台异步执行数据清洗、存储和 CII 计算。通过 GET /upload/{upload_id}/status 轮询处理状态。",
        "operationId": "upload_standard_csv_upload_vessel__vessel_id__standard_post",
        "parameters": [
          {
            "name": "vessel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "船舶 ID",
              "title": "Vessel Id"
            },
            "description": "船舶 ID"
          },
          {
            "name": "pitch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "螺旋桨螺距（用于滑失比计算）",
              "default": 6.058,
              "title": "Pitch"
            },
            "description": "螺旋桨螺距（用于滑失比计算）"
          },
          {
            "name": "vessel_capacity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "CII 计算吨位（载重吨或总吨），不提供则跳过 CII 计算",
              "title": "Vessel Capacity"
            },
            "description": "CII 计算吨位（载重吨或总吨），不提供则跳过 CII 计算"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/DataBody_upload_standard_csv_upload_vessel__vessel_id__standard_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseModel_UploadAccepted_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8003",
          "description": "Data 遥测数据服务"
        }
      ]
    },
    "/upload/vessel/{vessel_id}/history": {
      "get": {
        "summary": "获取船舶上传历史",
        "description": "按创建时间倒序返回该船舶的上传记录列表，支持分页。",
        "operationId": "get_upload_history_upload_vessel__vessel_id__history_get",
        "parameters": [
          {
            "name": "vessel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "船舶 ID",
              "title": "Vessel Id"
            },
            "description": "船舶 ID"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "每页条数",
              "default": 10,
              "title": "Limit"
            },
            "description": "每页条数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseModel_list_UploadStatusSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8003",
          "description": "Data 遥测数据服务"
        }
      ]
    },
    "/upload/{upload_id}/status": {
      "get": {
        "summary": "查询上传处理状态",
        "description": "返回单条上传记录，status 字段为 pending | processing | done | failed。",
        "operationId": "get_upload_status_upload__upload_id__status_get",
        "parameters": [
          {
            "name": "upload_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "上传记录 ID",
              "title": "Upload Id"
            },
            "description": "上传记录 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseModel_UploadStatusSchema_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8003",
          "description": "Data 遥测数据服务"
        }
      ]
    },
    "/daily/vessel/{vessel_id}": {
      "get": {
        "summary": "获取日均数据",
        "description": "返回 DuckDB 中存储的日均聚合数据（含 CII），按日期倒序分页。",
        "operationId": "get_daily_data_daily_vessel__vessel_id__get",
        "parameters": [
          {
            "name": "vessel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "船舶 ID",
              "title": "Vessel Id"
            },
            "description": "船舶 ID"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "偏移量",
              "default": 0,
              "title": "Offset"
            },
            "description": "偏移量"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 365,
              "minimum": 1,
              "description": "每页条数",
              "default": 30,
              "title": "Limit"
            },
            "description": "每页条数"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponseModel_list_DailyDataSchema__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataHTTPValidationError"
                }
              }
            }
          }
        }
      },
      "servers": [
        {
          "url": "http://localhost:8003",
          "description": "Data 遥测数据服务"
        }
      ]
    }
  },
  "components": {
    "schemas": {
...
      "DataBody_upload_standard_csv_upload_vessel__vessel_id__standard_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "CSV 标准数据文件"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_standard_csv_upload_vessel__vessel_id__standard_post"
      },
      "DataDailyDataSchema": {
        "properties": {
          "vessel_id": {
            "type": "integer",
            "title": "Vessel Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "speed_ground": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Ground"
          },
          "speed_water": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Water"
          },
          "me_shaft_power": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Me Shaft Power"
          },
          "me_rpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Me Rpm"
          },
          "me_fuel_consumption_nmile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Me Fuel Consumption Nmile"
          },
          "me_hfo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Me Hfo Act Cons"
          },
          "me_mgo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Me Mgo Act Cons"
          },
          "blr_hfo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blr Hfo Act Cons"
          },
          "blr_mgo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blr Mgo Act Cons"
          },
          "dg_hfo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dg Hfo Act Cons"
          },
          "dg_mgo_act_cons": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dg Mgo Act Cons"
          },
          "slip_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slip Ratio"
          },
          "draft": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft"
          },
          "wind_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wind Speed"
          },
          "cii_temp": {
            "type": "number",
            "title": "Cii Temp"
          },
          "cii": {
            "type": "number",
            "title": "Cii"
          }
        },
        "type": "object",
        "required": [
          "vessel_id",
          "date",
          "speed_ground",
          "speed_water",
          "me_shaft_power",
          "me_rpm",
          "me_fuel_consumption_nmile",
          "me_hfo_act_cons",
          "me_mgo_act_cons",
          "blr_hfo_act_cons",
          "blr_mgo_act_cons",
          "dg_hfo_act_cons",
          "dg_mgo_act_cons",
          "slip_ratio",
          "draft",
          "wind_speed",
          "cii_temp",
          "cii"
        ],
        "title": "DailyDataSchema"
      },
      "DataHTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/DataValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "DataResponseModel_UploadAccepted_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/DataUploadAccepted"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[UploadAccepted]"
      },
      "DataResponseModel_UploadStatusSchema_": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/DataUploadStatusSchema"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[UploadStatusSchema]"
      },
      "DataResponseModel_list_DailyDataSchema__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DataDailyDataSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[list[DailyDataSchema]]"
      },
      "DataResponseModel_list_UploadStatusSchema__": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "default": 200
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/DataUploadStatusSchema"
            },
            "type": "array",
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "success"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ResponseModel[list[UploadStatusSchema]]"
      },
      "DataUploadAccepted": {
        "properties": {
          "upload_id": {
            "type": "integer",
            "title": "Upload Id"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "上传成功，后台处理中"
          }
        },
        "type": "object",
        "required": [
          "upload_id"
        ],
        "title": "UploadAccepted"
      },
      "DataUploadStatusSchema": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vessel_id": {
            "type": "integer",
            "title": "Vessel Id"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "date_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Start"
          },
          "date_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date End"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "vessel_id",
          "file_path",
          "date_start",
          "date_end",
          "status",
          "error_message",
          "created_at",
          "completed_at"
        ],
        "title": "UploadStatusSchema"
      },
      "DataValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}